Database connection and CategoryForm

+++ src/main/java/com/tiwilli/bechosystem/model/entities/Category.java
@@ -5,22 +5,22 @@ import java.util.Objects;
 
 public class Category implements Serializable {
 
-    private Long id;
+    private Integer id;
     private String name;
 
     public Category() {
     }
 
-    public Category(Long id, String name) {
+    public Category(Integer id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
